+++ Modules/services/Controller/ServicesinvoiceprojectController.php
@@ -128,6 +128,10 @@ class ServicesinvoiceprojectController extends InvoiceAbstractController {
             $modelInvoiceItem->editItemContent($id_space, $id_items[0]["id"], $content, $total_ht);
             $modelInvoice->setTotal($id_space, $id_invoice, $total_ht);
             $modelInvoice->setDiscount($id_space, $id_invoice, $discount);
+
+            $_SESSION['flash'] = InvoicesTranslator::InvoiceHasBeenSaved($lang);
+            $_SESSION['flashClass'] = 'success';
+
             Events::send([
                 "action" => Events::ACTION_INVOICE_EDIT,
                 "space" => ["id" => intval($id_space)],
@@ -208,7 +212,12 @@ class ServicesinvoiceprojectController extends InvoiceAbstractController {
                 $itemQuantities[] = $data[1];
                 $itemPrices[] = $data[2];
                 $itemQuantityTypes[] = $modelSeTypes->getType($modelServices->getItemType($id_space, $data[0]));
-                $total += floatval($data[1]) * floatval($data[2]);
+                if (is_numeric($data[1]) && is_numeric($data[2])) {
+                    $total += $data[1] * $data[2];
+                } else {
+                    $_SESSION['flash'] = InvoicesTranslator::NonNumericValue($lang);
+                    $_SESSION['flashClass'] = 'danger';
+                }   
                 if (count($data) == 4) {
                     $itemComments[] = $data[3];
                 } else {